Improvements to migration logic

+++ beta/MXWeather.sh
@@ -2,20 +2,23 @@
 
 # Migrate v3 to v4 functionality
 
-# Checks if there is more than 1 file in the data folder (indicates new install or existing)
-if [ "$(ls -A /opt/CumulusMX/data/ | wc -l)" -gt 1 ]; then 
-# Checks for datav3 folder. If it doesn't exist then creates it and copies v3 files to it and makes a backup. 
-    if [ ! -f "/opt/CumulusMX/config/.migrated" ]; then 
-      # Copy Cumulus.ini to root
-      cp -f /opt/CumulusMX/config/Cumulus.ini /opt/CumulusMX/
-      # Backup data files
-      mkdir -p /opt/CumulusMX/backup/datav3
-      cp -R /opt/CumulusMX/data/* /opt/CumulusMX/backup/datav3
-      # Copy data files to datav3 for migration
-      mkdir /opt/CumulusMX/datav3
-      cp -R /opt/CumulusMX/data/* /opt/CumulusMX/datav3
-      # Run migration script
-      expect <<EOF
+# Enables migration if the environment variable is set
+if [ "$MIGRATE" ]; then
+echo "Migration enabled. Starting migration..."
+  # Checks to see if data has already been migrated and skips if it has. 
+  if [ ! -f "/opt/CumulusMX/config/.migrated" ]; then 
+    # Backup Cumulus.ini file
+    cp -R /opt/CumulusMX/config/Cumulus.ini /opt/CumulusMX/config/Cumulus-v3.ini.bak
+    # Copy Cumulus.ini to root
+    cp -f /opt/CumulusMX/config/Cumulus.ini /opt/CumulusMX/
+    # Backup data files
+    mkdir -p /opt/CumulusMX/backup/datav3
+    cp -R /opt/CumulusMX/data/* /opt/CumulusMX/backup/datav3
+    # Copy data files to datav3 for migration
+    mkdir /opt/CumulusMX/datav3
+    cp -R /opt/CumulusMX/data/* /opt/CumulusMX/datav3
+    # Run migration script
+    expect <<EOF
 spawn dotnet MigrateData3to4.dll
 expect "Press a Enter to continue, or Ctrl-C to exit"
 send "\r"
@@ -23,17 +26,16 @@ expect "Press Enter to exit"
 send "\r"
 expect eof
 EOF
-      # Leave a file to indicate the migration has been completed
-      touch /opt/CumulusMX/config/.migrated
-      # Copy UniqueID file to config folder
-      cp -f /opt/CumulusMX/UniqueId.txt /opt/CumulusMX/config/
-    else 
-      # If the .migrated file already exists it skips the migration. 
-      echo "Migration already completed."
-    fi
+    # Leave a file to indicate the migration has been completed
+    touch /opt/CumulusMX/config/.migrated
+    # Copy UniqueID file to config folder
+    cp -f /opt/CumulusMX/UniqueId.txt /opt/CumulusMX/config/
+  else 
+    # If the .migrated file already exists it skips the migration. 
+    echo "Migration already completed... Skipping migration."
+  fi
 else
-  # No data detected so there's nothing to migrate. 
-    echo "No data detected. Skipping migration."
+ echo "Migration not enabled... Skipping migration."
 fi
 
 # Start NGINX web server
@@ -54,8 +56,12 @@ term_handler() {
     kill -SIGTERM "$pid"
     wait "$pid"
     sleep 2
-    cp -f /opt/CumulusMX/Cumulus.ini /opt/CumulusMX/config/
-    cp -f /opt/CumulusMX/UniqueId.txt /opt/CumulusMX/config/
+    if [ -f "/opt/CumulusMX/config/Cumulus.ini" ]; then
+      cp -f /opt/CumulusMX/Cumulus.ini /opt/CumulusMX/config/
+    fi
+    if [ -f "/opt/CumulusMX/UniqueId.txt" ]; then
+      cp -f /opt/CumulusMX/UniqueId.txt /opt/CumulusMX/config/
+    fi
   fi
   exit 143; # 128 + 15 -- SIGTERM
 }
